SmartExpenseOCR/requirements.txt
fastapi
uvicorn
paddleocr
paddlepaddle
opencv-python
pillow
python-multipart
